@@ -2,14 +2,34 @@ import 'package:flutter/material.dart';
 
 // Define a custom Form widget.
 class EventForm extends StatefulWidget {
+  final Function addEvent;
+
+  EventForm(this.addEvent);
   @override
-  EventFormState createState() {
-    return EventFormState();
-  }
+  EventFormState createState() => EventFormState();
 }
 
 class EventFormState extends State<EventForm> {
   final _formKey = GlobalKey<FormState>();
+
+  final _nameController = TextEditingController();
+  final _descriptionController = TextEditingController();
+  DateTimeRange _dateRange = new DateTimeRange(
+    start: DateTime.now(),
+    end: DateTime.now().add(
+      Duration(days: 5),
+    ),
+  );
+
+  final _nameFocus = FocusNode();
+  final _descriptionFocus = FocusNode();
+
+  @override
+  void dispose() {
+    _nameController.dispose();
+    _descriptionController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -27,7 +47,7 @@ class EventFormState extends State<EventForm> {
           ),
           ElevatedButton(
             onPressed: () {
-              if (_formKey.currentState.validate()) {
+              if (_formKey.currentState!.validate()) {
                 ScaffoldMessenger.of(context)
                     .showSnackBar(SnackBar(content: Text('Processing Data')));
               }
